made changes on the ProfileImage component, also move unnecessary files from (tabs)"
