+++ app/(tabs)/shift-details-page/[id].tsx
@@ -5,7 +5,7 @@ import {ThemedView} from "@/components/ThemedView";
 import {weekdays, months} from "moment";
 import Ionicons from "@expo/vector-icons/Ionicons";
 import {Colors} from '@/constants/Colors'
-import {useShiftNavigation} from "@/app/shift-navigation";
+import { useShiftNavigation } from "@/hooks/shift-navigation";
 
 export default function ShiftDetailsPage () {
     const item = useLocalSearchParams();
