+++ components/ProfileImage.tsx
@@ -1,12 +1,21 @@
-import React, { useState } from "react";
+import type React from "react";
+import { useState } from "react";
 import { Image, StyleSheet, View, Pressable, Alert } from "react-native";
 import * as ImagePicker from "expo-image-picker";
 
 type ProfileImageProps = {
-  initialImageSource: string; // Initial image URL or local path
+  initialImageSource?: string; // Initial image URL or local path (optional)
+  width?: number;
+  height?: number;
+  borderRadius?: number;
 };
 
-export const ProfileImage: React.FC<ProfileImageProps> = ({ initialImageSource }) => {
+export const ProfileImage: React.FC<ProfileImageProps> = ({
+  initialImageSource,
+  width = 125, // Default width
+  height = 125, // Default height
+  borderRadius = 62.5, // Default to make it circular for the default size
+}) => {
   const defaultImage = "https://via.placeholder.com/125"; // Replace with your default image URL
   const [imageSource, setImageSource] = useState<string>(initialImageSource || defaultImage);
 
@@ -24,34 +33,32 @@ export const ProfileImage: React.FC<ProfileImageProps> = ({ initialImageSource }
   };
 
   const pickImage = async () => {
-    // Verify permissions before proceeding
     const hasPermission = await verifyPermissions();
     if (!hasPermission) {
       return;
     }
 
-    // Open the image picker
     const result = await ImagePicker.launchImageLibraryAsync({
       mediaTypes: ImagePicker.MediaTypeOptions.Images,
       allowsEditing: true,
-      aspect: [1, 1], // Crop the image to a square
-      quality: 1, // Highest quality
+      aspect: [1, 1],
+      quality: 1,
     });
 
     if (!result.canceled) {
-      setImageSource(result.assets[0].uri); // Update the image source with the selected image
+      setImageSource(result.assets[0].uri);
     }
   };
 
   return (
     <Pressable onPress={pickImage}>
-      <View style={styles.container}>
+      <View style={[styles.container, { width, height, borderRadius }]}>
         <Image
           source={{ uri: imageSource }}
-          style={styles.image}
+          style={[styles.image, { width: '100%', height: '100%' }]}
           onError={() => {
             Alert.alert("Image Load Error", "Failed to load the image. Reverting to default image.");
-            setImageSource(defaultImage); // Fallback to default image on error
+            setImageSource(defaultImage);
           }}
         />
       </View>
@@ -61,19 +68,14 @@ export const ProfileImage: React.FC<ProfileImageProps> = ({ initialImageSource }
 
 const styles = StyleSheet.create({
   container: {
-    width: 125,
-    height: 125,
-    borderRadius: 62.5,
-    overflow: "hidden",
     justifyContent: "center",
     alignItems: "center",
     marginBottom: 16,
     borderWidth: 2,
     borderColor: "#ccc",
+    overflow: "hidden", // Keep this for borderRadius to work
   },
   image: {
-    width: "100%",
-    height: "100%",
     resizeMode: "cover",
   },
 });
